requirements.txt
pandas>=1.5.0
numpy>=1.21.0
jupyter>=1.0.0
streamlit>=1.20.0
pymongo>=4.0.0
python-dotenv>=1.0.0
plotly>=5.10.0
fpdf>=1.7.2